chore(web): extract Sync panel to /theguard/connections/sync
Phase 2 Chunk A panel 2 of 4. Moves the "Sync status" tab out of the
Guard settings tabs into a dedicated route +
features/guard/connections/sync/Panel.tsx.

Settings page shim now redirects both ?tab=notifications and ?tab=sync
to their new URLs. No API, schema, or UX behavior changes — copy-move
only; toolCoverage fetch + handleResync moved into the panel so the
settings page no longer loads sync data it never displays.
